reorganized/styling/comments
